update16tensorboard.py
Hello, 
I downloaded this project to learn more about tensor board. However this file would not run because not all data was being sent to the GPU. I followed along  with your video of creating this file and found a point where it failed on my system. My proposed change is sending the example data to the GPU on line 83, this solved the problem for me. Hope this is helpful. Thanks for your time. 
Brendan